+++ api-v1/openapi.yaml
@@ -8,7 +8,10 @@ info:
   version: 1.1.0
 servers:
   - description: 九州大学休講情報 API v1
-    url: https://kyukou-kyudai.herokuapp.com/api/1
+    url: https://kyukou-kyudai.herokuapp.com/{basePath}
+    variables:
+      basePath:
+        default: api/1
 tags:
   - name: event
     description: Events API
