+++ routes/calendar.js
@@ -7,7 +7,7 @@ const vobject = require('vobject');
 
 const router = express.Router();
 
-const eventsAPI = require('../api1').events;
+const eventsAPI = require('../api-v1').events;
 
 router.get('/', (req, res) => {
   res.render('calendar', { site });
